requirements.txt
# Euclid Q1 ViT Galaxy Morphology — Dependencies
# Install: pip install -r requirements.txt

# Core
torch>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=2.0
pyarrow                   # Parquet support for pandas

# Data & ML
scikit-learn>=1.3
Pillow>=9.0

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Config
pyyaml>=6.0

# Downloading
requests>=2.28

# Transformers (for ViT, DINOv2, Swin, etc.)
transformers>=4.35
timm>=0.9                 # PyTorch Image Models (ConvNeXt, Swin-V2)

# Interpretability
grad-cam>=1.4             # GradCAM for CNN models

# Evaluation
scipy>=1.10               # Statistical tests

# Notebook
jupyter>=1.0
ipywidgets>=8.0
tqdm>=4.65
